+++ tests/rotation_generator.py
